renaming files, slightly more readable tests
